requirements.txt
schedule
requests
bs4
fuzzywuzzy
mysql-connector-python